feat: sprint 2 step 2 code review (move modal escape and add uuid)
