+++ src/services/reducers/index.jsx
@@ -11,6 +11,7 @@ import {
   GET_ORDER_NUMBER_FAILED,
   REPLACE_INGREDIENTS
 } from '../actions';
+import { v4 as uuidv4 } from 'uuid';
 
 const initialState = {
   ingredients: [],
@@ -57,6 +58,7 @@ export const getIngredientsReducer = (state = initialState, action) => {
       }
     }
     case ADD_INGREDIENT_TO_CONSTRUCTOR: {
+      action.draggedIngredient.uuid = uuidv4();
       return {
         ...state,
         constructorIngredients: [...state.constructorIngredients, action.draggedIngredient]
